requirements.txt
accelerate==0.27.2
datasets==2.17.1
jsonargparse==4.27.5
loralib==0.1.2
numpy==1.21.6
pandas==2.0.0
peft==0.10.0

scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.13.2

spacy==3.5.4
torch==2.2.1
tqdm==4.64.1
transformers==4.43.1
bitsandbytes==0.44.0
trl==0.8.3
wandb==0.16.4
python-levenshtein
